[REF] General refactoring/fixing
