+++ docs/source/esg_matching.engine.builders.rst
@@ -1,14 +1,16 @@
-esg\_matching.db\_engine.builders package
+esg\_matching.engine.builders
 =========================================
 
-Submodules
-----------
+**Package path:** esg\_matching.engine.builders
+
+.. automodule:: esg_matching.engine.builders
 
 esg\_matching.engine.builders.column\_builder module
 --------------------------------------------------------
 
 .. automodule:: esg_matching.engine.builders.column_builder
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
@@ -17,6 +19,7 @@ esg\_matching.engine.builders.delete\_builder module
 
 .. automodule:: esg_matching.engine.builders.delete_builder
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
@@ -25,6 +28,7 @@ esg\_matching.engine.builders.select\_builder module
 
 .. automodule:: esg_matching.engine.builders.select_builder
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
@@ -33,6 +37,7 @@ esg\_matching.engine.builders.sql\_condition\_builder module
 
 .. automodule:: esg_matching.engine.builders.sql_condition_builder
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
@@ -41,13 +46,7 @@ esg\_matching.engine.builders.table\_builder module
 
 .. automodule:: esg_matching.engine.builders.table_builder
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
-Module contents
----------------
-
-.. automodule:: esg_matching.engine.builders
-   :members:
-   :undoc-members:
-   :show-inheritance:
