+++ docs/source/esg_matching.engine.connectors.rst
@@ -1,29 +1,33 @@
-esg\_matching.db\_engine.engines package
+esg\_matching.db\_engine.engines
 ========================================
 
-Submodules
-----------
+**Package path:** esg\_matching.engine.connectors
 
-esg\_matching.engine.engines.connector module
+.. automodule:: esg_matching.engine.connectors
+
+esg\_matching.engine.connectors.base\_connector
 -------------------------------------------------
 
-.. automodule:: esg_matching.engine.engines.connector
+.. automodule:: esg_matching.engine.connectors.base_connector
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
-esg\_matching.engine.engines.connector\_oracle module
+esg\_matching.engine.connectors.oracle
 ---------------------------------------------------------
 
-.. automodule:: esg_matching.engine.engines.connector_oracle
+.. automodule:: esg_matching.engine.connectors.oracle
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
 
-esg\_matching.engine.engines.connector\_sql\_lite module
+esg\_matching.engine.connectors.sql\_lite
 ------------------------------------------------------------
 
-.. automodule:: esg_matching.engine.engines.connector_sql_lite
+.. automodule:: esg_matching.engine.connectors.sql_lite
    :members:
+   :private-members:
    :undoc-members:
    :show-inheritance:
