+++ esg_matching/engine/__init__.py
@@ -1,0 +1,9 @@
+""" The **esg_matching.engine** package provides the following subpackages to work with different databases
+    using a common interface and making the details of database connection and specificity, transparent for
+    the client application:
+
+    - **builders:** provides modules that helps to construct database structures (columns, tables and SQL statements).
+    - **connectors:** provides modules to connect to a specific database type (e.g. oracle and sql lite)
+    - **sql:** provides modules to execute DML and DQL statements
+
+"""
