+++ esg_matching/engine/builders/table_builder.py
@@ -6,7 +6,7 @@ from sqlalchemy import MetaData, Table
 
 # Import internal modules
 from esg_matching.exceptions import exceptions_db_engine, exceptions_builders
-from esg_matching.engine.connectors.base_connector import DbConnector
+from esg_matching.engine.connectors.base import DbConnector
 from esg_matching.engine.builders.column_builder import ColumnBuilder
