+++ esg_matching/engine/connectors/__init__.py
@@ -1,0 +1,4 @@
+"""
+    The **esg_matching.engine.connectors** package provides modules to connect to a specific database.
+    Currently, the library supports Oracle and SQL Lite.
+"""
