+++ esg_matching/engine/connectors/oracle.py
@@ -6,7 +6,7 @@ import sqlalchemy as sa
 from sqlalchemy.dialects import oracle
 
 # Import internal modules
-from esg_matching.engine.connectors.base_connector import DbConnector
+from esg_matching.engine.connectors.base import DbConnector
 from esg_matching.exceptions import exceptions_db_engine
