+++ esg_matching/engine/sql/__init__.py
@@ -1,0 +1,3 @@
+"""
+    The **esg_matching.engine.sql** package provides modules to execute DML and DQL statements.
+"""
