+++ esg_matching/engine/sql/dml.py
@@ -4,7 +4,7 @@
 import sqlalchemy as sa
 
 # Import internal modules
-from esg_matching.engine.connectors.base_connector import DbConnector
+from esg_matching.engine.connectors.base import DbConnector
 
 
 class DmlManager:
